release 3.1.43 source code for python

+++ huaweicloud-sdk-functiongraph/huaweicloudsdkfunctiongraph/v2/model/month_used.py
@@ -18,7 +18,7 @@ class MonthUsed:
 
     openapi_types = {
         'date': 'str',
-        'value': 'int'
+        'value': 'float'
     }
 
     attribute_map = {
@@ -34,7 +34,7 @@ class MonthUsed:
         :param date: 日期
         :type date: str
         :param value: 使用量
-        :type value: int
+        :type value: float
         """
         
         
@@ -77,7 +77,7 @@ class MonthUsed:
         使用量
 
         :return: The value of this MonthUsed.
-        :rtype: int
+        :rtype: float
         """
         return self._value
 
@@ -88,7 +88,7 @@ class MonthUsed:
         使用量
 
         :param value: The value of this MonthUsed.
-        :type value: int
+        :type value: float
         """
         self._value = value
